Split sfrp_behavior into 4 components - integrated pmSplitabble.
sfrp_cachedpull, sfrp_cachedincreasingpartialtimefunction, and
sfrp_increasingpartialtimefunction were created.

The pmSplittable function was removed because this behavior is now
integrated into 'Behavior' directly.

The constructor of Behavior was removed and replaced with a 'fromValuePullFunc'
static function. This was done to emphasize that this is a lower-level call that
generally shouldn't be used.

Tests, documentation, and examples were also added.